Add test coverage for roundtrip serialization
