--- test/TesterInternal/SerializerTests/RoundTripSerializerTests.cs
@@ -1,18 +1,0 @@
-﻿using Orleans;
-using TestExtensions;
-using UnitTests.GrainInterfaces;
-using Xunit;
-
-namespace UnitTests.SerializerTests
-{
-    public class RoundTripSerializerTests : HostedTestClusterEnsureDefaultStarted
-    {
-        [Fact, TestCategory("Functional"), TestCategory("Serialization")]
-        public void Serialize_TestMethodResultEnum()
-        {
-            var grain = GrainClient.GrainFactory.GetGrain<IEnumResultGrain>(GetRandomGrainId());
-            CampaignEnemyTestType result = grain.GetEnemyType().Result;
-            Assert.Equal(CampaignEnemyTestType.Enemy2, result); //Enum return value wasn't transmitted properly
-        }
-    }
-}
